Don't recompute IR children
Profiling benchmarks revealed that we spend non-neglible amount of time
building IR preorder which in turns recomputes IR nodes' children.
IR is immutable so there is not need to recompute that data for every
pass.
Note that the change uses `lazy val` because benchmarks weren't showing
much improvements when allocating children on creation (with `val`). We
could probably improve the situation by using a poor version of `lazy
val` that does not employ synchronization.