Use simple type name as Jackson type id
Using a fully qualified name for type identifier does not play well when
dealing with different classloaders (classes are not recognized as
equal/subtypes).
There appears to be a number of casses loosely related issues on that
subject. Gave up on adding a test cases to simulate the problem but I
could no longer reproduce it with LS/runtime on a live project.