Avoid out-of-context interrupted lock exceptions (#10095)
This change is a re-do of old PR https://github.com/enso-org/enso/pull/8270 that has been abandoned a while ago.

The change attempts to address to case of random `InterruptedExceptions` that can result when a task is interrupted while waiting on a lock. In comparison to the original PR, this change introduces `withXYZLock` methods with a `Callable` parameter. As a result we ensure correct acquisition/release semantics, logging and handling of interruptions.

# Important Notes
https://github.com/enso-org/enso/pull/8270 was too much hassle to rebase against latest develop.