Introduce ImportExportScope proxy
`ImportExportScope` serves as a proxy to the underlying module's scope.
Can't reference the scope directly at the point when it is built because
building of some modules might still be in progress.

This also allowed to get rid off the horrible `withTypes` method that
would create a temporary module scope.