Drop built/build check for ModuleScope.Builder
Simplified the whole approach by avoiding calls to `built` and `build`
to determine the state of `ModuleScope.Builder`.
Instead, `ModuleScope.Builder` is only passed whenever a registration is
necessary. Since builder is accessed in various places, while it may
still, in theory, be constructor, this change introduces a simple,
short-lived, proxy that delegates to accessors and satisfies the
`ModuleScope` API.