Soft retries when awaiting runtime responses
A quick and dirty workaround for slow processing in a similar spirit to
PR #9858.
Long compilation of stdlib holds a write compilation lock, while opening
the file needs to set module sources and requires read compilation lock
and file lock. The expectation was that setting module sources is
instantaneous except not, because of locks.

This PR adds soft retries.