Amortize the startup cost of some components
Building Engine, Context, ApplicationConfig and Ydoc was a adding a
rather large delay during the initial startup step as all of those were
blocking operations.
Moving all of those to the resource initialization step hopes to
amortize some of that cost since it can be done in parallel.
Had to add a `ComponentSupervisor` (open for a different name
suggestion) to ensure that such delayed components are properly closed
on shutdown.