Ag grid boolean inference turned off for mixed type column (#10683)
If one value in a column is a boolean ag-grid renders all values in the column as a checkbox when cellDataTypes is true. This becomes a problem when getting a row of mixed type, so in this situation cellDataType will be set to false.

Bug:
![row-mixed-type-bug-problem](https://github.com/user-attachments/assets/5fcdd96d-8422-4370-b71e-94181ac42e2e)

Fix:
![row-mixed-type-bug-fix](https://github.com/user-attachments/assets/9ad18d1b-1b55-4351-a3ee-57a390b91fc4)

(cherry picked from commit c4e8ae143e4c1b7e0b69aa83daa608cf5ba02270)