Workaround for Electron sanbox issues on Ubuntu
Disabling sanbox, while a non-secure option, appears to be currently the
only option to workaround Electron issues on latest Ubuntu (24.04).
When passed to Enso it is treated as a regular argument, preventing the
start of the product.
This change treats `--no-sandbox` specially by ignoring all arguments
that precede it (and including itself).