Set JNA boot dir explicitly
Alternative approach to deal with Windows/Mac failures on CI where
`jnidispatch` native library somehow could not be identified. Linux
worked just fine with the previous approach.

This change enhances custom classloader to identify the desired library.
That way we can provide the path to it, as specified in the
implementation of JNA.