Ensure logging does not report spurious failures
... and a bunch of other small improvements.

Previously, when shutting down a project in a normal way, user would be
spammed with
```
java.io.EOFException
	at java.base/java.io.ObjectInputStream$BlockDataInputStream.peekByte(ObjectInputStream.java:3232)
	at java.base/java.io.ObjectInputStream.readObject0(ObjectInputStream.java:1713)
	at java.base/java.io.ObjectInputStream.readObject(ObjectInputStream.java:540)
	at java.base/java.io.ObjectInputStream.readObject(ObjectInputStream.java:498)
	at org.enso.logging.service.logback.SocketLoggingNode.run(SocketLoggingNode.java:63)
	at java.base/java.lang.Thread.run(Thread.java:1583)
[DEBUG] [...] [org.enso.logging.service.logback.SocketLoggingNode] Caught java.io.EOFException closing connection.
```
That's because LoggingServer was unaware of LanguageServer being
stopped, or what client/project were actually connected with a socket.
This change ensures that no shutdown errors are being reported when in
shutdown mode.

In order to associate individual logging socket with a project, we make
use of Mapped Diagnostic Context (MDC) properties. Except that the info
was being lost during serialization. Had to provide a custom
`SocketAppender` that preserves MDC when event is being reported.

This has an additional advantage that MDC will be possible to be used
when logging events remotely and provide sufficient context for
indexing/search.

Added retries to `DummyRepository` as download tests were frequently
reporting false positives.

Added a single test to `NativeLibraryFinderTest` to investigate MacOS
failures.