Ensure project-manager's logs are centrally collected
Minor fix to bring back project-manager's logs.

Previous, default, configuration meant that logs for project-manager
ended up in a separate file from the rest of the logs.
Language server logs are being sent to the logging server by default,
so should project-manager's ones as well.

Fixed NPE in `SocketLoggingNode` when no `project.id` is present in
loggers, as is the case with logs coming from project-manager.