Add right click context menu to graph editor. (#12601)
Fixes #8616
Context menu that opens when right clicking on graph background. The context menu will not show up when right-click was combined as a drag and used as zoom action.

The actions are provided with additional context that presents the menu position when available. That way actions can change behavior depending on whether or not they are called from context menu. This behavior is currently used in "Add Component" action.

<img width="290" alt="image" src="https://github.com/user-attachments/assets/1a6a0f1c-4a4a-4ea7-9686-e598923b5335" />