Module index invalidation triggers analysis
Previous fix for #12859 was helpful, but not sufficient.
It revealed an inherent race-condition between
`InvalidateModulesIndexJob` and `AnalyzeModuleInScopeJob`, which could
lead to silent discarding of suggestions info for non-Main modules. It
was entirely time-related and could result in no widgets on startup.

This change ensures that any invalidation of modules' indexes always
triggers necessary compilation. The latter in turn would trigger
analysis of cancelled modules.