Java std-base is not calling Enso during init
When Java code is calling Enso methods via `invokeMember` in the static
initializer and `ThreadManager` happens to interrupt it, we may end up
with a broken initialization of the class (manifested as
`NoClassDefFoundError`).

This change delays initilization of `ReloadDetector` until it is really
needed.

Also rewrote #12930 - we really don't need all methods of
`java.io.Console` in Native Image.