Invalidate RHS if name of binding changes
If name in the expression binding is modified it may be the case that
some other expression using it will become and not invalidate its value.
This is especially visible when flipping the order of assignments and
making one dependent on another.
Since `ChangesetBuilder` operates solely on syntactic text edits, there
is currently no other way than invalidating a complete RHS of a modified
binding, including any self arguments inside it.

The unit test in `ChangesetBuilderTest` illustrates the problem.
This can later result in lack of cache invalidation and stale
visualizations.

Closes #12957.