Generalize Telemetry Appender to send all logs
This change refactors existing Telemetry logic, so that we now have
OpenSearch and Telemetry remote logging, both using _mostly_ the same
logic.
The split comes from the fact that OpenSearch appender requires a bit
more tweaks to its configuration. For example, we want to avoid an
infinite loop of logging when connection failures occur.