More efficient visualization upsert via retries
Evaluation of a visualization occasionally requires write compilation
lock to load/compile involved modules. This leads to common lock
contention situations when in the current implementation.

Upgrading from a read compilation lock to a write one leads to deadlock
and is not supported by Reentrant locks anyway. Requesting write
compilation lock always is also not acceptable because it impacts
performance.

The current solution uses retries:
- locations that trigger compilation in visualization evaluation are
  well known
- we first attempt to evaluate visualizations without a compilation and
  read lock only
- if the latter fails, we acquire write lock and retry