Ensure visualization upsert triggers execution when necessary
`ExecuteJob` needs to run whenever there was no observable present for
the requested expression ID. Had to bring back cache invalidation logic
to ensure that one can attach visualization to any expression. This was
problematic for situations when requested expression was part of a
larger one which result has already been cached.