Ensure module sources are initialized
If Module sources where loaded from a disk but not (yet) sent by a GUI
we had a potential race-condition in the execution.
A safety check that was retrieving sources of a method to-be-called
could have module's sources not yet initialized and ignoring the fact
that they had been loaded from a disk.

The most problematic situation was to reproduce the problem which only
happened on
1) first load of a project
2) no caches are yet available
3) NI build

The three conditions were not yet sufficient to have the "Missing
Main.main" error to happen every time. Had to introduce an artifical
sleep in `textFile/open` to create the necessary race-condition to
appear every time.

We could potentially remove the src condition alltogether, as it is now
covered in other places.