Ensure Node reset execution environment on exit
ProbeNode wrappers are created for individual nodes to support
instrumentation. It appears though that a wrapper created for `onEnter`
notification may not be the same wrapper calling `onReturnValue`
notification for the same underlying node.
This is problematic as there is a potentially `originalExecutionEnvironment`
maintained between the two calls. As a result, a one-off execution
environment is never reset and any future execution continues to use the
provided environment (typically `live`) which is incorrect.

This change maintains the history of set node-specific execution
environments in a separate map of the instrumentation factory.
Additionally `getExecutionEnvironment` becomes a  destructive, non-idempotent
operation, as it should.