Fix `proxyRefs` bug
Fixes the "Help" button in the component menu; it was failing with a runtime exception.

The bug was caused by type-unsafety in Vue's `proxyRefs`: The function returns
an object with all fields writable, regardless of the types of refs in the
fields of the object provided as its argument. Created a more-strictly-typed
re-export that types fields proxied to readonly refs as readonly fields. Added
a linter rule forbidding direct usage of Vue's `proxyRefs`.