Support multiple execute expression requests
The original implementation limited the number of "one-off" expressions
to 1. In the event when a single request takes long time, all the other
requests were being stalled. Eventually, not picked up jobs would
trigger cancellations which would cause restart loops.

There is still a problem with some visualizations showing `ERR` for
tables. It appears to be a GUI issue.