Component documentation in Markdown 2/2 (#13203)
Second part of #12983

This removes the doc parser, addresses the remaining review comments from the first PR, and fixes some minor issues in documentation breadcrumbs.

This one should be merged **after** migration to the markdown in the stdlib docs.

https://github.com/user-attachments/assets/78ea34e6-fd4e-4ea0-8ea7-99bc5dd60d68

# Important Notes
Integration tests are failing because we need to regenerate the mock suggestions file. It will be easier when the docs in the stdlib are updated to the new format.

(cherry picked from commit 46c07100774b6ff7b0372cdccfe8829cae8ca8bc)