Work around a bug in Veaury.
The Veaury `onBeforeUnmount` hook fails if it is run when the `onMounted` hook
was not. Vue may run the unmounting hooks without running `onMounted` when a
component is unmounted before its `onMounted` hook has been run. In
`reactComponent`, patch the component created by `applyPureVueInReact` to skip
running the cleanup logic if it was never actually mounted. Add a lint rule to
ensure the buggy function is never used directly.

Fixes #13688.