Exponential backoff in executeExpression request
Modified hardcoded timeout when waiting for data in `executeExpression`
request. Most errors in table visualizations were caused because
computation was simply taking more time than allowed with a (short) timeout.
`executeExpression` was never going to give any guarantees when it comes
to returning data withing the specified timeframe. This is because it is
dependent on the execution of the program and any possible delays
related to it.

Rather than blindly increasing the timeout, added a retry with
exponenential backoff. This fixes the frequent `ERR` problems in table
visualizations.