Stabilize Table_Visualization tests
If data quality metrics are still pending for distinct count, it would
return `Nothing`. This in turn meant that our tests in `Table_Spec`
could be flaky as `is_using_multi_filter` would return `False` in that
case. This was more visible in our native image builds as things were
faster.
Not sure why the behaviour was changed in #13583.