Cache all arguments
In real-life, larger workflows we see re-executions taking place due to
visualizations being attached to more and more arguments.

The situation is expected as non-self arguments were not cached and are
being recomputed all the time. But impact on performance, especially
during startup, is very much visible.
This change caches all arguments. While small, the change has
significant impact on performance during startup as we appear to avoid
most of the numerous re-executions.