Replace synchronization blocks in polyglot Java initialization
If 2 threads are trying to initialize polyglot Java
- Thread 1 acquires the lock on the object and proceeds with
 initialization
- Thread 2 waits on the lock on the object

Sadly, Thread 1, while keeping the lock hits the TruffleSafepoint
barrier, waiting for all other threads to poll at least once.
Thread 2 will never call TruffleSafepoint.poll() as it is waiting.

This change replaces synchronization block with a lightweight access to
a sempahore. Single cell semaphore lock is attempted with a `tryAcquire`
call until successful. `acquire` must not be used as it would lead to
the problem described above again.
Busy polling for a sempahore is being used. It seems acceptable as
creating a language is not a resource-intensive process and based on
local testing proceeds quickly with little wasted resource.